Fix bug with wrong double cointing. Change all double to BigDecimal.

+++ src/main/java/pl/damianlegutko/fprecrutation/user/UserService.java
@@ -2,9 +2,11 @@ package pl.damianlegutko.fprecrutation.user;
 
 import pl.damianlegutko.fprecrutation.user.api.UserDTO;
 
+import java.math.BigDecimal;
+
 interface UserService {
     UserDTO findUserByUsername(String username);
     void saveUser(UserDTO user);
-    void giveMoneyToUser(String userName, double moneyAmount);
-    void takeMoneyFromUser(String userName, double moneyAmount);
+    void giveMoneyToUser(String userName, BigDecimal moneyAmount);
+    void takeMoneyFromUser(String userName, BigDecimal moneyAmount);
 }
