+++ src/main/java/pl/damianlegutko/fprecrutation/user/api/UserController.java
@@ -10,6 +10,8 @@ import pl.damianlegutko.fprecrutation.user.exceptions.UserException;
 import pl.damianlegutko.fprecrutation.user.exceptions.UserHaveNotEnoughMoneyException;
 import pl.damianlegutko.fprecrutation.user.exceptions.UserNotExistsException;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/api/user")
 class UserController {
@@ -38,14 +40,14 @@ class UserController {
     //TODO temporary REST
     @PostMapping("/addCash/{userName}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    void increaseUserCash(@PathVariable String userName, @RequestParam double amount) {
+    void increaseUserCash(@PathVariable String userName, @RequestParam BigDecimal amount) {
         user2service.giveMoneyToUser(userName, amount);
     }
 
     //TODO temporary REST
     @PostMapping("/subtractCash/{userName}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    void decreaseUserCash(@PathVariable String userName, @RequestParam double amount) {
+    void decreaseUserCash(@PathVariable String userName, @RequestParam BigDecimal amount) {
         user2service.takeMoneyFromUser(userName, amount);
     }
